+++ gwcelery/tests/test_sentry.py
@@ -18,24 +18,22 @@ def netrcfile(monkeypatch, tmpdir):
     yield filename
 
 
-@patch('raven.contrib.celery.register_logger_signal', autospec=True)
-@patch('raven.contrib.celery.register_signal', autospec=True)
-@patch('raven.Client', autospec=True)
-def test_sentry_configure(mock_client, mock_register_signal,
-                          mock_register_logger_signal, netrcfile, caplog):
+@patch('sentry_sdk.init')
+@patch('sentry_sdk.integrations.celery.CeleryIntegration')
+def test_sentry_configure(mock_celery_integration, mock_sdk_init,
+                          netrcfile, caplog):
     caplog.set_level(logging.ERROR)
     sentry.configure()
     record, = caplog.records
     assert 'Disabling Sentry integration' in record.message
-    mock_client.assert_not_called()
-    mock_register_signal.assert_not_called()
-    mock_register_logger_signal.assert_not_called()
+    mock_celery_integration.assert_not_called()
+    mock_sdk_init.assert_not_called()
 
     scheme, netloc, *rest = urlparse(sentry.DSN)
     with open(netrcfile, 'w') as f:
         print('machine', netloc, 'login', 'foo', 'password', 'bar', file=f)
     sentry.configure()
-    dsn = urlunparse((scheme, 'foo:bar@' + netloc, *rest))
-    mock_client.assert_called_once_with(dsn)
-    mock_register_signal.assert_called_once()
-    mock_register_logger_signal.assert_called_once()
+    dsn = urlunparse((scheme, 'foo@' + netloc, *rest))
+    mock_celery_integration.assert_called_once_with()
+    mock_sdk_init.assert_called_once_with(
+        dsn, integrations=[mock_celery_integration.return_value])