+++ CHANGES.md
@@ -33,6 +33,9 @@
 -   Depend on ligo-gracedb >= 2.0.1 in order to pull in a bug fix for VOEvents
     with ProbHasNS or ProbHasRemnant set to 0.0.
 
+-   Use the `sentry-sdk` package instead of the deprecated `raven` package for
+    Sentry integration.
+
 ## 0.1.0 (2018-09-26)
 
 -   Separated the external GCN listening handlers into two: one that listens